ipconfig: fixed multiple issues with ip, subnetmask and netbios
specifically:
to determine which ip's and sm are changeable validate against
a registry key.  this eliminates the query to wmi which returns
ip which we cannot manage (specifically VIPs for clusters)

additionally cleaned up netbios logic.
its important to note, you cannot set netbios to dhcp if the
adapter is not on dhcp.  this is not a valid setting